Refactor Go client into multiple files
