+++ go_client/update.go
@@ -1,0 +1,117 @@
+package main
+
+import (
+	"archive/tar"
+	"compress/gzip"
+	"encoding/binary"
+	"fmt"
+	"io"
+	"net/http"
+	"os"
+	"path/filepath"
+	"strings"
+)
+
+func downloadGZ(url, dest string) error {
+	resp, err := http.Get(url)
+	if err != nil {
+		return err
+	}
+	defer resp.Body.Close()
+	if resp.StatusCode != http.StatusOK {
+		return fmt.Errorf("GET %s: %s", url, resp.Status)
+	}
+	gz, err := gzip.NewReader(resp.Body)
+	if err != nil {
+		return err
+	}
+	defer gz.Close()
+	tmp := dest + ".tmp"
+	f, err := os.Create(tmp)
+	if err != nil {
+		return err
+	}
+	if _, err := io.Copy(f, gz); err != nil {
+		f.Close()
+		return err
+	}
+	if err := f.Close(); err != nil {
+		return err
+	}
+	return os.Rename(tmp, dest)
+}
+
+func downloadTGZ(url, destDir string) error {
+	resp, err := http.Get(url)
+	if err != nil {
+		return err
+	}
+	defer resp.Body.Close()
+	if resp.StatusCode != http.StatusOK {
+		return fmt.Errorf("GET %s: %s", url, resp.Status)
+	}
+	gz, err := gzip.NewReader(resp.Body)
+	if err != nil {
+		return err
+	}
+	defer gz.Close()
+	tr := tar.NewReader(gz)
+	for {
+		hdr, err := tr.Next()
+		if err == io.EOF {
+			break
+		}
+		if err != nil {
+			return err
+		}
+		path := filepath.Join(destDir, hdr.Name)
+		switch hdr.Typeflag {
+		case tar.TypeDir:
+			if err := os.MkdirAll(path, 0755); err != nil {
+				return err
+			}
+		case tar.TypeReg:
+			if err := os.MkdirAll(filepath.Dir(path), 0755); err != nil {
+				return err
+			}
+			f, err := os.Create(path)
+			if err != nil {
+				return err
+			}
+			if _, err := io.Copy(f, tr); err != nil {
+				f.Close()
+				return err
+			}
+			if err := f.Close(); err != nil {
+				return err
+			}
+		}
+	}
+	return nil
+}
+
+func autoUpdate(resp []byte) error {
+	if len(resp) < 16 {
+		return fmt.Errorf("short response for update")
+	}
+	base := string(resp[16:])
+	if i := strings.IndexByte(base, 0); i >= 0 {
+		base = base[:i]
+	}
+	base = strings.TrimRight(base, "/")
+	clientVer := binary.BigEndian.Uint32(resp[4:8])
+	fmt.Printf("Client version: %v\n", clientVer)
+	imgVer := binary.BigEndian.Uint32(resp[8:12])
+	sndVer := binary.BigEndian.Uint32(resp[12:16])
+	imgURL := fmt.Sprintf("%s/data/CL_Images.%d.gz", base, imgVer>>8)
+	sndURL := fmt.Sprintf("%s/data/CL_Sounds.%d.gz", base, sndVer>>8)
+	fmt.Println("downloading", imgURL)
+	if err := downloadGZ(imgURL, "CL_Images"); err != nil {
+		return err
+	}
+	fmt.Println("downloading", sndURL)
+	if err := downloadGZ(sndURL, "CL_Sounds"); err != nil {
+		return err
+	}
+	return nil
+}
